+++ src/Common/FileCache_fwd.h
@@ -9,8 +9,8 @@ static constexpr int REMOTE_FS_OBJECTS_CACHE_DEFAULT_MAX_FILE_SEGMENT_SIZE = 100
 static constexpr int REMOTE_FS_OBJECTS_CACHE_DEFAULT_MAX_ELEMENTS = 1024 * 1024;
 static constexpr int REMOTE_FS_OBJECTS_CACHE_ENABLE_HITS_THRESHOLD = 0;
 
-class IFileCache;
-using FileCachePtr = std::shared_ptr<IFileCache>;
+class FileCache;
+using FileCachePtr = std::shared_ptr<FileCache>;
 
 struct FileCacheSettings;
