+++ src/Common/IFileCachePriority.h
@@ -5,7 +5,8 @@
 #include <mutex>
 #include <Core/Types.h>
 #include <Common/Exception.h>
-#include <Common/IFileCache.h>
+#include <Common/FileCache.h>
+#include <Common/FileCacheType.h>
 
 namespace DB
 {
@@ -15,6 +16,8 @@ namespace ErrorCodes
     extern const int NOT_IMPLEMENTED;
 }
 
+class FileCache;
+
 class IFileCachePriority;
 using FileCachePriorityPtr = std::shared_ptr<IFileCachePriority>;
 
@@ -22,20 +25,14 @@ using FileCachePriorityPtr = std::shared_ptr<IFileCachePriority>;
 class IFileCachePriority
 {
 public:
-    struct Key
-    {
-        UInt128 key;
-        String toString() const;
-
-        Key() = default;
-        explicit Key(const UInt128 & key_) : key(key_) { }
-
-        bool operator==(const Key & other) const { return key == other.key; }
-    };
-
     class IIterator;
     friend class IIterator;
-    using Iterator = std::shared_ptr<IIterator>;
+
+    using ReadIterator = std::shared_ptr<const IIterator>;
+    using WriteIterator = std::shared_ptr<IIterator>;
+
+    friend class FileCache;
+    using Key = FileCacheKey;
 
     struct FileCacheRecord
     {
@@ -56,30 +53,6 @@ public:
     public:
         virtual ~IIterator() = default;
 
-        virtual void next() { throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support next() for IIterator."); }
-
-        virtual bool valid() const { throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support valid() for IIterator."); }
-
-        ///  Mark a cache record as recently used, it will update the priority
-        /// of the cache record according to different cache algorithms.
-        virtual void use(std::lock_guard<std::mutex> &)
-        {
-            throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support use() for IIterator.");
-        }
-
-        /// Deletes an existing cached record.
-        virtual void remove(std::lock_guard<std::mutex> &)
-        {
-            throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support remove() for IIterator.");
-        }
-
-        virtual Iterator getSnapshot() { throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support getSnapshot() for IIterator."); }
-
-        virtual void incrementSize(size_t, std::lock_guard<std::mutex> &)
-        {
-            throw Exception(ErrorCodes::NOT_IMPLEMENTED, "Not support incrementSize() for IIterator.");
-        }
-
         virtual Key key() const = 0;
 
         virtual size_t offset() const = 0;
@@ -87,6 +60,23 @@ public:
         virtual size_t size() const = 0;
 
         virtual size_t hits() const = 0;
+
+        virtual void next() const = 0;
+
+        virtual bool valid() const = 0;
+
+        ///  Mark a cache record as recently used, it will update the priority
+        /// of the cache record according to different cache algorithms.
+        virtual void use(std::lock_guard<std::mutex> &) = 0;
+
+        /// Deletes an existing cached record.
+        virtual void remove(std::lock_guard<std::mutex> &) = 0;
+
+        virtual WriteIterator getWriteIterator() const = 0;
+
+        virtual void incrementSize(size_t, std::lock_guard<std::mutex> &) = 0;
+
+        virtual void seekToLowestPriority() const = 0;
     };
 
 public:
@@ -94,7 +84,7 @@ public:
 
     /// Add a cache record that did not exist before, and throw a
     /// logical exception if the cache block already exists.
-    virtual Iterator add(const Key & key, size_t offset, size_t size, std::lock_guard<std::mutex> & cache_lock) = 0;
+    virtual WriteIterator add(const Key & key, size_t offset, size_t size, std::lock_guard<std::mutex> & cache_lock) = 0;
 
     /// Query whether a cache record exists. If it exists, return true. If not, return false.
     virtual bool contains(const Key & key, size_t offset, std::lock_guard<std::mutex> & cache_lock) = 0;
@@ -103,7 +93,7 @@ public:
 
     /// Returns an iterator pointing to the lowest priority cached record.
     /// We can traverse all cached records through the iterator's next().
-    virtual Iterator getNewIterator(std::lock_guard<std::mutex> & cache_lock) = 0;
+    virtual ReadIterator getNewIterator(std::lock_guard<std::mutex> & cache_lock) = 0;
 
     virtual size_t getElementsNum(std::lock_guard<std::mutex> & cache_lock) const = 0;
