+++ src/Disks/IO/CachedReadBufferFromRemoteFS.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Common/IFileCache.h>
+#include <Common/FileCache.h>
 #include <IO/SeekableReadBuffer.h>
 #include <IO/WriteBufferFromFile.h>
 #include <IO/ReadSettings.h>
@@ -81,7 +81,7 @@ private:
     bool writeCache(char * data, size_t size, size_t offset, FileSegment & file_segment);
 
     Poco::Logger * log;
-    IFileCache::Key cache_key;
+    FileCache::Key cache_key;
     String remote_fs_object_path;
     FileCachePtr cache;
     ReadSettings settings;
@@ -128,7 +128,7 @@ private:
     CurrentMetrics::Increment metric_increment{CurrentMetrics::FilesystemCacheReadBuffers};
     ProfileEvents::Counters current_file_segment_counters;
 
-    IFileCache::QueryContextHolder query_context_holder;
+    FileCache::QueryContextHolder query_context_holder;
 
     bool is_persistent;
 };
