+++ src/Disks/ObjectStorages/DiskObjectStorage.cpp
@@ -10,7 +10,8 @@
 #include <Common/quoteString.h>
 #include <Common/logger_useful.h>
 #include <Common/filesystemHelpers.h>
-#include <Common/IFileCache.h>
+#include <Disks/IO/ThreadPoolRemoteFSReader.h>
+#include <Common/FileCache.h>
 #include <Disks/ObjectStorages/DiskObjectStorageRemoteMetadataRestoreHelper.h>
 #include <Disks/ObjectStorages/DiskObjectStorageTransaction.h>
 #include <Disks/FakeDiskTransaction.h>
