+++ src/Disks/ObjectStorages/IObjectStorage.h
@@ -16,6 +16,7 @@
 #include <Disks/IO/AsynchronousReadIndirectBufferFromRemoteFS.h>
 #include <Disks/ObjectStorages/StoredObject.h>
 #include <Common/ThreadPool.h>
+#include <Common/FileCache.h>
 #include <Disks/WriteMode.h>
