+++ src/Interpreters/InterpreterDescribeCacheQuery.cpp
@@ -6,7 +6,7 @@
 #include <DataTypes/DataTypeString.h>
 #include <Storages/ColumnsDescription.h>
 #include <Common/FileCacheFactory.h>
-#include <Common/IFileCache.h>
+#include <Common/FileCache.h>
 #include <Access/Common/AccessFlags.h>
 #include <Core/Block.h>
